41: create a Flask blueprint called `api_bp` (= extract the API route(-handling function)s into their own module)
